move purchase order list and view to their own components
moved view and list to their own components so they can be re-used
